feat(navbar): adapt miniprogram's capsule length change feature (#2922)
* feat(navbar): adapt miniprogram's capsule length change feature

wechat add a new feature of displaying the provider of the miniprogram on the capsule. When it
displays, the capsule enxtends above the content of the navbar, this commit adapts this new feature
, hides the content of the navbar when it is covered by the capsule

"feat #2802"

* feat(navbar): dapt miniprogram's capsule length change feature

wechat add a new feature of displaying the provider of the miniprogram on the capsule. When it
displays, the capsule enxtends above the content of the navbar, this commit adapts this new feature
, hides the content of the navbar when it is covered by the capsule

"feat #2802"

* feat(navbar): update snap

* feat(navbar): use getRect adaptor

---------

Co-authored-by: jarmywang <awangjianjun9@gmail.com>